getFunctionSignatures.js Switched to version that does not rely on truffle.
