+++ package.json
@@ -17,7 +17,7 @@
     "test": "node scripts/check_for_clashing.js && bash scripts/test.sh",
     "testOnTestnet": "./node_modules/.bin/truffle test --network testNet",
     "testNoUpgradeProxy": "export CONTRACTS_NO_UPGRADE_PROXY=true && node scripts/check_for_clashing.js && bash scripts/test.sh",
-    "getFunctionSignatures": "./node_modules/.bin/truffle exec scripts/getFunctionSignatures.js",
+    "getFunctionSignatures": "node scripts/getFunctionSignatures.js",
     "feedTestAccountsOnTestnet": "./node_modules/.bin/truffle exec scripts/feedTestAccounts.js --network testNet",
     "testDev": "bash scripts/test.sh test/KeyGenHistory.js test/mockContracts/BlockRewardHbbftCoinsMock.sol  test/mockContracts/RandomHbbftMock.sol test/mockContracts/ValidatorSetHbbftMock.sol test/mockContracts/StakingHbbftCoinsMock.sol"
   },
